feat: :sparkles: Implementação de controle de permissões e hierarquia de usuários nas operações de resposta e tópico

+++ src/main/java/br/com/forum_hub/domain/resposta/RespostaService.java
@@ -4,7 +4,9 @@ import br.com.forum_hub.domain.topico.Status;
 import br.com.forum_hub.domain.topico.TopicoService;
 import br.com.forum_hub.domain.usuario.Usuario;
 import br.com.forum_hub.infra.exception.RegraDeNegocioException;
+import br.com.forum_hub.infra.seguranca.HierarquiaService;
 import jakarta.transaction.Transactional;
+
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,10 +15,12 @@ import java.util.List;
 public class RespostaService {
     private final RespostaRepository repository;
     private final TopicoService topicoService;
+    private final HierarquiaService hierarquiaService;
 
-    public RespostaService(RespostaRepository repository, TopicoService topicoService) {
+    public RespostaService(RespostaRepository repository, TopicoService topicoService, HierarquiaService hierarquiaService) {
         this.repository = repository;
         this.topicoService = topicoService;
+        this.hierarquiaService = hierarquiaService;
     }
 
     @Transactional
@@ -38,8 +42,12 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta atualizar(DadosAtualizacaoResposta dados) {
+    public Resposta atualizar(DadosAtualizacaoResposta dados, Usuario logado) {
         var resposta = buscarPeloId(dados.id());
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getAutor(), "ROLE_MODERADOR"))
+            throw new RegraDeNegocioException("Você não pode atualizar essa resposta!");
+
         return resposta.atualizarInformacoes(dados);
     }
 
@@ -48,10 +56,14 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta marcarComoSolucao(Long id) {
+    public Resposta marcarComoSolucao(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
 
         var topico = resposta.getTopico();
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_INSTRUTOR"))
+            throw new RegraDeNegocioException("Você não pode marcar essa resposta como solução!");
+
         if(topico.getStatus() == Status.RESOLVIDO)
             throw new RegraDeNegocioException("O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
 
@@ -60,9 +72,12 @@ public class RespostaService {
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
         var topico = resposta.getTopico();
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getAutor(), "ROLE_MODERADOR"))
+            throw new RegraDeNegocioException("Você não pode marcar excluir essa resposta!");
 
         repository.deleteById(id);
