+++ src/main/java/br/com/forum_hub/domain/topico/TopicoService.java
@@ -3,6 +3,7 @@ package br.com.forum_hub.domain.topico;
 import br.com.forum_hub.domain.curso.CursoService;
 import br.com.forum_hub.domain.usuario.Usuario;
 import br.com.forum_hub.infra.exception.RegraDeNegocioException;
+import br.com.forum_hub.infra.seguranca.HierarquiaService;
 import jakarta.transaction.Transactional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -14,10 +15,12 @@ public class TopicoService {
 
     private final TopicoRepository repository;
     private final CursoService cursoService;
+    private final HierarquiaService hierarquiaService;
 
-    public TopicoService(TopicoRepository repository, CursoService cursoService) {
+    public TopicoService(TopicoRepository repository, CursoService cursoService, HierarquiaService hierarquiaService) {
         this.repository = repository;
         this.cursoService = cursoService;
+        this.hierarquiaService = hierarquiaService;
     }
 
     @Transactional
@@ -26,7 +29,9 @@ public class TopicoService {
         var topico = new Topico(dados, curso, autor);
         return repository.save(topico);
     }
-    public Page<DadosListagemTopico> listar(String categoria, Long idCurso, Boolean semResposta, Boolean solucionados, Pageable paginacao) {
+
+    public Page<DadosListagemTopico> listar(String categoria, Long idCurso, Boolean semResposta, Boolean solucionados,
+            Pageable paginacao) {
         Specification<Topico> spec = Specification.where(TopicoSpecification.estaAberto())
                 .and(TopicoSpecification.temCategoria(categoria))
                 .and(TopicoSpecification.temCursoId(idCurso))
@@ -38,15 +43,23 @@ public class TopicoService {
     }
 
     @Transactional
-    public Topico atualizar(DadosAtualizacaoTopico dados) {
+    public Topico atualizar(DadosAtualizacaoTopico dados, Usuario logado) {
         var topico = buscarPeloId(dados.id());
         var curso = cursoService.buscarPeloId(dados.cursoId());
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR"))
+            throw new RegraDeNegocioException("Você não pode atualizar esse tópico!");
+
         return topico.atualizarInformacoes(dados, curso);
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) {
         var topico = buscarPeloId(id);
+
+        if (hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR"))
+            throw new RegraDeNegocioException("Você não pode apagar esse tópico!");
+
         if (topico.getStatus() == Status.NAO_RESPONDIDO)
             repository.deleteById(id);
         else
